+++ packages/fe-fpm-writer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/fe-fpm-writer
 
+## 0.31.22
+
+### Patch Changes
+
+-   Updated dependencies [e93797a]
+    -   @sap-ux/project-access@1.28.9
+    -   @sap-ux/fiori-annotation-api@0.3.3
+
 ## 0.31.21
 
 ### Patch Changes